requirements.txt
hindsight-client>=0.5.0
litellm>=1.40.0
streamlit>=1.35.0
python-dotenv>=1.0.0
groq>=0.9.0
